+++ middleware2/src/main/java/expense/data/ExpenseRepository.java
@@ -2,14 +2,12 @@ package expense.data;
 
 import expense.model.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
-public interface ExpenseRepository extends JpaRepository<Expense,Long> {
+public interface ExpenseRepository extends JpaRepository<Expense,Integer> {
 
     List<Expense> findAll();
